requirements.txt
transformers==4.30.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
datasets==2.13.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
nltk==3.8.1
evaluate==0.4.0
rouge-score==0.1.2
sacrebleu==2.3.1
tqdm==4.65.0
farasapy
arabert
accelerate>=0.20.1 